add react-router to example, separate balance stuff into its own component-page
